Add indentation Offset to the first line of the formatted document
Fix computation of indentation offset when indentation Kind is Tabs and
add test for it.
Move Indentation_Offset from Format_Options_Type to
Identation_Options_Type. The indentation offset now models smart tabs
where spaces are used both for indentation and alignemnt, or, tabs
are used for indentation and spaces for alignment.
This is needed for range formatting with tabs. For instance:
```
declare
   Obj_Foo : constant Type_Foo :=
     Value_Foo;
```
To range format Value_Foo we need an offset of 1 tab and 2 spaces.