Allow formatting of No_Document
Assertions that check that no No_Documents are present were removed.
The reason is because one could create a Document_Type object as
follows:
`Document : Document_Type;`
This is equivalent to No_Document.
To disallow this, we would have to make the type Document_Type limited,
which is not desirable.
Therefore, we should allow `No_Document` and simply skip any operations
on it.